PWN-7913 - Comment fixes
