+++ app/src/main/java/org/p2p/wallet/history/ui/historylist/HistoryListView.kt
@@ -19,8 +19,6 @@ import org.p2p.uikit.utils.toPx
 import org.p2p.wallet.common.ui.recycler.EndlessScrollListener
 import org.p2p.wallet.common.ui.recycler.PagingState
 import org.p2p.wallet.databinding.LayoutHistoryListBinding
-import org.p2p.wallet.history.ui.historylist.HistoryListViewContract.HistoryListViewType
-import org.p2p.wallet.history.ui.historylist.HistoryListViewContract.View.HistoryListViewClickListener
 import org.p2p.wallet.history.ui.model.HistoryItem
 import org.p2p.wallet.history.ui.token.adapter.HistoryAdapter
 import org.p2p.wallet.utils.unsafeLazy
