+++ app/src/main/java/org/p2p/wallet/history/ui/historylist/HistoryListViewClickListener.kt
@@ -1,0 +1,7 @@
+package org.p2p.wallet.history.ui.historylist
+
+interface HistoryListViewClickListener {
+    fun onTransactionClicked(transactionId: String)
+    fun onSellTransactionClicked(transactionId: String)
+    fun onUserSendLinksClicked()
+}
